Update getAlbums method in index.js
According to the documentation, the parameter of the getAlbums() method should be provided as an object in order to work; not an array.  I modified the getAlbums() method to send the params object within an array to the getAlbumsMany() method.  I've tested the change and all is working as expected.