Update PHAssetsService.m
We're almost there.  I updated the reversed startIndex and max startIndex and endIndex values based on assetCount to correctly match up with the required assets.

This properly fetches the assets form newest to oldest when reversed is true, but these values that are returned need to be reversed again in order to display properly.  I hope that makes sense! I can provide more details if needed.

Also, I think this reversed display should be the default since it mimics the way photos are loaded for the native Photos app.