Fixed sorting of assets.
This fix should do the trick!  By default, assets will be loaded from newest to oldest.  If the reverse prop is true, oldest assets will be loaded first.  I've tested in my app and all is running smoothly!

Now, we need to decide if the "reverse" prop is descriptive enough.  Perhaps renaming this prop to "reverseAssetSort" would be more descriptive since this prop reverses the sort on the assets being loaded?  Let me know your thoughts and suggestions.  Glad this is working as expected!